Generate a data source name (DSN) for the GDAL MEM driver

An array in memory can be referenced by a GDAL data source.

mem(
  x,
  extent = NULL,
  projection = "",
  dimension = NULL,
  PIXELOFFSET = NULL,
  LINEOFFSET = NULL,
  BANDOFFSET = NULL
)

Arguments

x

an R array, must be of numeric type (integer is converted to double)

extent

optional extent of the data in x,y c(xmin, xmax, ymin, ymax)

projection

projection string (optional, sets the SPATIALREFERENCE of the MEM driver since GDAL 3.7)

dimension

size in pixels, ncol, nrow

PIXELOFFSET

pixel offset

LINEOFFSET

line offset

BANDOFFSET

band offset

Value

character string, a DSN for use by GDAL

Details

This DSN will only work in R, and is only for use with GDAL read and query tools (so terra, sf, gdalcubes, vapour, etc.).

Examples

m <- matrix(as.integer(c(0L, 0, 0, 1)), 5L, 4L)
mem(m)
#> [1] "MEM:::DATAPOINTER=\"94855832884240\",PIXELS=5,LINES=4,BANDS=1,DATATYPE=Float64,GEOTRANSFORM=0/1/0/4/0/-1,PIXELOFFSET=0,LINEOFFSET=0,BANDOFFSET=160"
mem(volcano)
#> [1] "MEM:::DATAPOINTER=\"94855930984672\",PIXELS=87,LINES=61,BANDS=1,DATATYPE=Float64,GEOTRANSFORM=0/1/0/61/0/-1,PIXELOFFSET=0,LINEOFFSET=0,BANDOFFSET=42456"